+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/javascript/__tests__/serializeJSComponent.test.ts
@@ -223,6 +223,7 @@ describe('serializeJSComponent', () => {
     it('serializes component with Flow property types', () => {
       // given
       const component: ComponentImplementationInfo = getImplementation('IntegrationComboFlow');
+      console.log('FLOW', component);
       const expectedMetadata: ComponentMetadata = {
         defaultExported: true,
         name: 'IntegrationComboFlow',